Fix id, update and delete.
Previously, the lib did not set the model.id attribute when populating a
collection from firebase.  This meant that the models could not be identified
when `child_changed` or `child_removed` events came down the pipe.

This commit introduces a configurable `idAttribute`, which is populated with
the firebase ference name.  Implementations then need to set the `idAttrubute`
of their model class to use this as the source for the `model.id` value.  I've
updated the readme accordingly but it might be worth providing a `FirebaseModel`
class that does this automatically, or generally considering the best
implementation.